Move challenge scripts from utils to controllers
